Fixed KatBox autosizing
Atleast one of the KatBox subsite comics were using considerably large images, while also setting in the element's style it's size. The Reader script decided to interpret both to create a horizontally squished image. So basically added in some styling rules to fix this. Unfortunately had to use !important, as the site's style is embedded directly into the element instead of CSS.