DHS Comix Supported
DHS Comix Comics now supported, albeit a little messily. Only so much can be done for a site that has random pages with multiple images, as well as very little little ID or Class for elements. Atleast I got it down to two entries, instead of having to put in a entry for every single sub comic. This is what happens when the only identifiers for next and back are image file names, and one of the comics has it's own set of file names for the nav buttons. Ah well, it works.

Probably will be my last change for a while.