Let me tell you about Homestuck
Fixed DMFA/MissMab, Curtailed, and initial Homestuck work

When Homestuck changed to the new domain, it also got a revamped site. it no longer requires the multiple entries for detection depending on various things like which act we're in. This is good.

It however decided to be a pain to setup for whatever reason. Took several hours but I managed to get it finally.

It properly shows the image, the Pesterchat, and even the nav link. But since the link wasn't automatically switching to being #next, I used some JS to make it like the image and navigating by clicking left or right on it's box.

The other comics on the site, like Problem Sleuth aren't setup yet. That's on the way.


Also AppleGeeks is being a big pain in getting to work. Two days (on and off) have been wasted trying to get that to work.

Also think I forgot several commits ago to mention I fixed some Kevin and Kell aligning problems, particularly on smaller screens like phones.

Current List of To Fix:

Yosh Saga
Artificial Incident
Dilbert
Lackadaisy Cats
Vicki Fox
Three Panel Soul (minor fix due to the front page)
The various sites that have been moving off of Katbox that aren't ending up on Kemono Cafe
Two Kinds