Update webcomic_reader.user.js
Added Yosh Saga and Artificial Incident
Fixed Vicki Fox

Previous comitt: Added Fullscreen. Primarily for Mobile Users to clear up realestate. 
Examples: 
https://i.imgur.com/09x8xMJ.png
https://i.imgur.com/ThrFZfg.png

Also note for mobile users: It may be best to either enable "Request Desktop site" in the browser, or enable Fit-to-Screen with shrink width to fit enabled. Maybe on newer phones it isn't an issue, but on a Samsung S4 I've been testing this on, it likes to display pages zoomed in. And while in Fullscreen you can't zoom out.