Update webcomic_reader.user.js
Setup default appearances for settings buttons. as well as setting a width to the settings dialog.
Need to remember later to set the default font and size later.

Also minor addition to Kevin and Kell, listing the page's title as an extra. Which appears to have been in the original setup. Contemplating if I should add in the comments section to the extra section as well. Might test it to see how much of a problem it causes since it's using Disqus.

Added in the ability to browse FA's galleries if they weren't on a comic page. Though if a comic page that uses the built in comic nav system hits the queue it will start going into that comic's pathing. Might have to look into some JS to get it to go "if I started in viewing Gallery images, stay in gallery navigation". It's a minor issue that might be a little annoying.

Also added the native gallery queue to the extra. Will work on styling for it later.

....still no luck on AppleGeeks. What I have SHOULD be working from what I can tell. It's just... not....