Update webcomic_reader.user.js
Well the Kevin and Kell extra change from last comitt was a bust. Because the title also contains a bit of javascript to detect how many comments are made on that page, it causes the image to have a lag time when switching pages.

Also found out that the "Disable on this site" button also only appears on mobile, maybe even just for the mobile version of greasemonkey. Will look into this more with tampermonkey as well. Would like to style these buttons, but they only have text and nothing else to ID them. And since CSS doesn't have a selector for what a tag contains between it... I even have to wonder if Javascript can touch the buttons since they should technically be added only after the script has ran.

I thought of several solutions, but they have their down sides
1. Default style for all buttons.
-Problem: This affects all buttons on a site
2. Use nth last child of type css selector to select the last button
-Problem: If this is ran outside of mobile, this will affect the last two buttons on a page.
3. Appended a hidden button towards the end of the page via script
-Problem: Too much effort for a minor thing.

These three are not ideal.