Update webcomic_reader.user.js
It's messy looking if the script first loads on it, but this should fix the one(?) page in Girl Genius that has multiple images. Gave it additional leeway in the event they do that again, but with more images (up to 10)

Same can be said about Bob and George from the previous revision.