Update webcomic_reader.user.js
Ok, previous fix did fix the problem. Had the function being called for every page cache. I fixed it in a not-so elegant manner though.. It's embedded in the check on whether the script will put in "enable this script" or "disable this script" checks for script commands.

Also this contains a minor fix for Homestuck. It'll make it so any pages with a movie will not be cached. It was trying to cache those pages which resulted in no video, as well as the need to disable the script to view said movie files.