fixed issue with squished images
It seemed randomly, but some sites would render with a squished or stretched image. Turns out, it's likely due to the site setting "100%" to max-width or max-height. This should fix that for most if not all sites. Just needed two lines to set the max sizes in the same section where it sets the width and height to be [native or scaled]