wcr_ext_navi_ctrls + Twokinds Extra Navigation Fix
Added a function that I've reused multiple times to save space.
By just having "js:	wcr_ext_navi_ctrls" in an entry, it will make it so that anything in the "Extra" section will be counted as navigable like the primary page. This will help somewhat on pages that have multiple images. Just have the extra images be caught by "extra:" and throw that in at the bottom of the comic's entry, and bam!

The reason for this is that the same function has been used multiple times in the past. It wasn't a dedicated function, thus it was code that was being repeated several times. Thus: single function that could be called anytime, presto!

This won't work for every instance, for an example Homestuck. Due to the way that one is setup in the capture, it's got it's own edited version of this function running.

Also twokinds' extra now is now navigable as well.