2020.07.27.160900 - Exit Fullscreen Fix
An outstanding issue with the fullscreen function was if you left fullscreen from any other method than the added button it would require you to press the button several times or reload the page. It should now detect when fullscreen has exited and change the varialbes as well as the text on the button to reflect it. So it should now be able to go back into fullscreen easily.