2020.08.03.152600 - Girl Genius&EGS extra captures
Redid Girl Genius's Extra capture. Didn't like how it didn't scroll. And after several attempts at creating a variation of the extra navigation function to do auto scrolling (scrollear() prevents this), I redid the capture setup. It resulted in less space used.

Also, during a re-read of EGS, I noticed that every so often Dan would have a link that went to a "alt" image. I put in a capture for that to display it (at 50% width to differentiate it in cases where he's showing off that he made an edit, and thus not-canon)