GG minor fix and Shivae comics
Minor fix (non-ideal) for the "latest page" of Girl Genius.

Added the rest of Shivae Comics, as well as a minor fix for Alien Dice.

Still have to look into several issues with these sites, primarily the lack of overflow scrolling, as well as the sites managing to color the setting window's buttons. Would take care of it now but I barely got the time for it right this minute.