2021.12.28.224300
Was not able to do a proper pull request nor rebase request to get this updated properly to the master branch's english translation patch.

Thankfully GitKraken was nice enough to use to help me go through the changes to mark what was what. Sadly couldn't submit this push via it due to some issues on my end. (should be resolved within a week or so)

It really saved me manually making copy and pastes for hours or even days.

Anyway, english translation should be up and running on this fork. Will soon be looking into applying it to the pull requester so we can get the anka main up to date with the changes I've made.