Setting up Personal Testing branch
This branch is for personal testing, primarily to see how the updateURL works on various platforms. If it works on mobile and what not, then it might be a branch I just use to push updates to for quick and ease updating on various devices for additional testing.

No support is going to be given on this branch.