Temp fix for Bob & George
So, I started re-reading through B&G's archives. Today, while I was progressing while in bed, Dave broke the comments section. So it is now using 5vw for the font size. Which means the font is extremely large. So this is a temp fix till he fixed that, if he notices.

Hopefully I'll remember to undo my fix when he fixes it.